[CrOS Cellular] Add ResetEuicc in eSIM uninstall handler.
This CL combines Reset EUICC operation with the uninstall pipeline so
that the same set of operations are performed in the reset case as
well. This fixes issue with services not always being cleared after
reset. This method is used in DeviceResetEuiccCommandJob class in
followup CL.

Bug: b/210697209
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3421792
Commit-Queue: Azeem Arshad <azeemarshad@chromium.org>
Cr-Commit-Position: refs/heads/main@{#966320}

+++ chromeos/network/cellular_esim_uninstall_handler.cc
@@ -6,6 +6,7 @@
 
 #include "base/containers/flat_set.h"
 #include "base/metrics/histogram_functions.h"
+#include "base/time/time.h"
 #include "chromeos/dbus/hermes/hermes_euicc_client.h"
 #include "chromeos/dbus/hermes/hermes_profile_client.h"
 #include "chromeos/network/cellular_esim_profile_handler.h"
@@ -22,19 +23,27 @@
 
 namespace chromeos {
 
+// static
+const base::TimeDelta CellularESimUninstallHandler::kNetworkListWaitTimeout =
+    base::Seconds(20);
+
 CellularESimUninstallHandler::UninstallRequest::UninstallRequest(
-    const std::string& iccid,
+    const absl::optional<std::string>& iccid,
     const absl::optional<dbus::ObjectPath>& esim_profile_path,
     const absl::optional<dbus::ObjectPath>& euicc_path,
+    bool reset_euicc,
     UninstallRequestCallback callback)
     : iccid(iccid),
       esim_profile_path(esim_profile_path),
       euicc_path(euicc_path),
+      reset_euicc(reset_euicc),
       callback(std::move(callback)) {}
 CellularESimUninstallHandler::UninstallRequest::~UninstallRequest() = default;
 
 CellularESimUninstallHandler::CellularESimUninstallHandler() = default;
-CellularESimUninstallHandler::~CellularESimUninstallHandler() = default;
+CellularESimUninstallHandler::~CellularESimUninstallHandler() {
+  OnShuttingDown();
+}
 
 void CellularESimUninstallHandler::Init(
     CellularInhibitor* cellular_inhibitor,
@@ -47,6 +56,8 @@ void CellularESimUninstallHandler::Init(
   network_configuration_handler_ = network_configuration_handler;
   network_connection_handler_ = network_connection_handler;
   network_state_handler_ = network_state_handler;
+
+  network_state_handler->AddObserver(this, FROM_HERE);
 }
 
 void CellularESimUninstallHandler::UninstallESim(
@@ -55,8 +66,37 @@ void CellularESimUninstallHandler::UninstallESim(
     const dbus::ObjectPath& euicc_path,
     UninstallRequestCallback callback) {
   uninstall_requests_.push_back(std::make_unique<UninstallRequest>(
-      iccid, esim_profile_path, euicc_path, std::move(callback)));
+      iccid, esim_profile_path, euicc_path, /*reset_euicc=*/false,
+      std::move(callback)));
   ProcessPendingUninstallRequests();
+}
+
+void CellularESimUninstallHandler::ResetEuiccMemory(
+    const dbus::ObjectPath& euicc_path,
+    UninstallRequestCallback callback) {
+  uninstall_requests_.push_back(std::make_unique<UninstallRequest>(
+      /*iccid=*/absl::nullopt, /*esim_profile_path=*/absl::nullopt, euicc_path,
+      /*reset_euicc=*/true, std::move(callback)));
+  ProcessPendingUninstallRequests();
+}
+
+void CellularESimUninstallHandler::NetworkListChanged() {
+  if (state_ != UninstallState::kWaitingForNetworkListUpdate) {
+    return;
+  }
+  // When removing multiple eSIM network services back to back after a Reset
+  // EUICC, uninstall handler will be in waiting state till next network list
+  // update before removing next configuration.
+  network_list_wait_timer_.Stop();
+  TransitionToUninstallState(UninstallState::kRemovingShillService);
+  AttemptRemoveShillService();
+}
+
+void CellularESimUninstallHandler::OnShuttingDown() {
+  if (network_state_handler_) {
+    network_state_handler_->RemoveObserver(this, FROM_HERE);
+    network_state_handler_ = nullptr;
+  }
 }
 
 void CellularESimUninstallHandler::ProcessPendingUninstallRequests() {
@@ -69,10 +109,10 @@ void CellularESimUninstallHandler::ProcessPendingUninstallRequests() {
   if (state_ != UninstallState::kIdle)
     return;
 
-  NET_LOG(EVENT) << "Starting eSIM uninstall. ICCID: "
-                 << GetIccidForCurrentRequest();
+  NET_LOG(EVENT) << "Starting eSIM uninstall for request "
+                 << *uninstall_requests_.front();
   TransitionToUninstallState(UninstallState::kCheckingNetworkState);
-  CheckNetworkState();
+  CheckActiveNetworkState();
 }
 
 void CellularESimUninstallHandler::TransitionToUninstallState(
@@ -90,8 +130,8 @@ void CellularESimUninstallHandler::CompleteCurrentRequest(
       "Network.Cellular.ESim.UninstallProfile.OperationResult", result);
 
   const bool success = result == UninstallESimResult::kSuccess;
-  NET_LOG(EVENT) << "Completed uninstall request for ICCID "
-                 << GetIccidForCurrentRequest() << ". Success = " << success;
+  NET_LOG(EVENT) << "Completed uninstall request for "
+                 << *uninstall_requests_.front() << ". Success = " << success;
   std::move(uninstall_requests_.front()->callback).Run(success);
   uninstall_requests_.pop_front();
 
@@ -99,15 +139,17 @@ void CellularESimUninstallHandler::CompleteCurrentRequest(
   ProcessPendingUninstallRequests();
 }
 
-const std::string& CellularESimUninstallHandler::GetIccidForCurrentRequest()
-    const {
-  return uninstall_requests_.front()->iccid;
-}
-
 const NetworkState*
 CellularESimUninstallHandler::GetNetworkStateForCurrentRequest() const {
+  absl::optional<std::string> current_request_iccid =
+      uninstall_requests_.front()->iccid;
+
+  if (!current_request_iccid) {
+    return nullptr;
+  }
+
   for (auto* const network : GetESimCellularNetworks()) {
-    if (network->iccid() == GetIccidForCurrentRequest()) {
+    if (network->iccid() == current_request_iccid) {
       return network;
     }
   }
@@ -115,28 +157,15 @@ CellularESimUninstallHandler::GetNetworkStateForCurrentRequest() const {
   return nullptr;
 }
 
-void CellularESimUninstallHandler::CheckNetworkState() {
+void CellularESimUninstallHandler::CheckActiveNetworkState() {
   DCHECK_EQ(state_, UninstallState::kCheckingNetworkState);
 
-  const NetworkState* network = GetNetworkStateForCurrentRequest();
-  if (!network) {
-    NET_LOG(ERROR) << "Unable to find eSIM network with ICCID "
-                   << GetIccidForCurrentRequest();
-    CompleteCurrentRequest(UninstallESimResult::kNetworkNotFound);
-    return;
-  }
-
-  // If there is no profile path in the request then this is a stale service.
-  // Skip directly to configuration removal.
-  if (!uninstall_requests_.front()->esim_profile_path) {
-    TransitionToUninstallState(UninstallState::kRemovingShillService);
-    AttemptRemoveShillService();
-    return;
-  }
+  const NetworkState* network = network_state_handler_->ActiveNetworkByType(
+      NetworkTypePattern::Cellular());
 
   // If the network is connected, disconnect it before we attempt to uninstall
-  // the associated profile.
-  if (network->IsConnectedState()) {
+  // eSIM profiles.
+  if (network && network->IsConnectedState()) {
     TransitionToUninstallState(UninstallState::kDisconnectingNetwork);
     AttemptNetworkDisconnect(network);
     return;
@@ -170,8 +199,8 @@ void CellularESimUninstallHandler::OnDisconnectFailure(
     std::unique_ptr<base::DictionaryValue> error_data) {
   DCHECK_EQ(state_, UninstallState::kDisconnectingNetwork);
 
-  NET_LOG(ERROR) << "Failed disconnecting network with ICCID "
-                 << GetIccidForCurrentRequest();
+  NET_LOG(ERROR) << "Failed disconnecting network for request "
+                 << *uninstall_requests_.front();
   CompleteCurrentRequest(UninstallESimResult::kDisconnectFailed);
 }
 
@@ -189,8 +218,8 @@ void CellularESimUninstallHandler::OnShillInhibit(
   DCHECK_EQ(state_, UninstallState::kInhibitingShill);
 
   if (!inhibit_lock) {
-    NET_LOG(ERROR) << "Error inhbiting Shill during uninstall for ICCID "
-                   << GetIccidForCurrentRequest();
+    NET_LOG(ERROR) << "Error inhbiting Shill during uninstall for request "
+                   << *uninstall_requests_.front();
     CompleteCurrentRequest(UninstallESimResult::kInhibitFailed);
     return;
   }
@@ -218,8 +247,9 @@ void CellularESimUninstallHandler::OnRefreshProfileListResult(
   DCHECK_EQ(state_, UninstallState::kRequestingInstalledProfiles);
 
   if (!inhibit_lock) {
-    NET_LOG(ERROR) << "Error refreshing profile list during uninstall for "
-                   << "ICCID " << GetIccidForCurrentRequest();
+    NET_LOG(ERROR)
+        << "Error refreshing profile list during uninstall for request "
+        << *uninstall_requests_.front();
     CompleteCurrentRequest(UninstallESimResult::kRefreshProfilesFailed);
     return;
   }
@@ -234,8 +264,20 @@ void CellularESimUninstallHandler::OnRefreshProfileListResult(
 
 void CellularESimUninstallHandler::AttemptDisableProfile() {
   DCHECK_EQ(state_, UninstallState::kDisablingProfile);
+  absl::optional<dbus::ObjectPath> esim_profile_path;
+  if (uninstall_requests_.front()->reset_euicc) {
+    esim_profile_path = GetEnabledCellularESimProfilePath();
+    // Skip disabling profile if there are no enabled profiles.
+    if (!esim_profile_path) {
+      TransitionToUninstallState(UninstallState::kUninstallingProfile);
+      AttemptUninstallProfile();
+      return;
+    }
+  } else {
+    esim_profile_path = uninstall_requests_.front()->esim_profile_path;
+  }
   HermesProfileClient::Get()->DisableCarrierProfile(
-      *uninstall_requests_.front()->esim_profile_path,
+      *esim_profile_path,
       base::BindOnce(&CellularESimUninstallHandler::OnDisableProfile,
                      weak_ptr_factory_.GetWeakPtr()));
 }
@@ -249,8 +291,8 @@ void CellularESimUninstallHandler::OnDisableProfile(
   bool success = status == HermesResponseStatus::kSuccess ||
                  status == HermesResponseStatus::kErrorAlreadyDisabled;
   if (!success) {
-    NET_LOG(ERROR) << "Failed to disable profile for ICCID "
-                   << GetIccidForCurrentRequest();
+    NET_LOG(ERROR) << "Failed to disable profile for request "
+                   << *uninstall_requests_.front();
     CompleteCurrentRequest(UninstallESimResult::kDisableProfileFailed);
     return;
   }
@@ -261,6 +303,15 @@ void CellularESimUninstallHandler::OnDisableProfile(
 
 void CellularESimUninstallHandler::AttemptUninstallProfile() {
   DCHECK_EQ(state_, UninstallState::kUninstallingProfile);
+
+  if (uninstall_requests_.front()->reset_euicc) {
+    HermesEuiccClient::Get()->ResetMemory(
+        *uninstall_requests_.front()->euicc_path,
+        hermes::euicc::ResetOptions::kDeleteOperationalProfiles,
+        base::BindOnce(&CellularESimUninstallHandler::OnUninstallProfile,
+                       weak_ptr_factory_.GetWeakPtr()));
+    return;
+  }
 
   HermesEuiccClient::Get()->UninstallProfile(
       *uninstall_requests_.front()->euicc_path,
@@ -273,11 +324,13 @@ void CellularESimUninstallHandler::OnUninstallProfile(
     HermesResponseStatus status) {
   DCHECK_EQ(state_, UninstallState::kUninstallingProfile);
 
-  hermes_metrics::LogUninstallProfileResult(status);
+  if (!uninstall_requests_.front()->reset_euicc) {
+    hermes_metrics::LogUninstallProfileResult(status);
+  }
 
   if (status != HermesResponseStatus::kSuccess) {
-    NET_LOG(ERROR) << "Failed to uninstall profile for ICCID "
-                   << GetIccidForCurrentRequest();
+    NET_LOG(ERROR) << "Failed to uninstall profile for request "
+                   << *uninstall_requests_.front();
     CompleteCurrentRequest(UninstallESimResult::kUninstallProfileFailed);
     return;
   }
@@ -289,22 +342,33 @@ void CellularESimUninstallHandler::OnUninstallProfile(
 void CellularESimUninstallHandler::AttemptRemoveShillService() {
   DCHECK_EQ(state_, UninstallState::kRemovingShillService);
 
-  const NetworkState* network = GetNetworkStateForCurrentRequest();
-  if (!network) {
-    NET_LOG(ERROR) << "Unable to find eSIM network with ICCID "
-                   << GetIccidForCurrentRequest();
-    CompleteCurrentRequest(UninstallESimResult::kRemoveServiceFailed);
-    return;
+  const NetworkState* network = nullptr;
+  if (uninstall_requests_.front()->reset_euicc) {
+    network = GetNextResetServiceToRemove();
+    if (!network) {
+      CompleteCurrentRequest(UninstallESimResult::kSuccess);
+      return;
+    }
+  } else {
+    network = GetNetworkStateForCurrentRequest();
+    if (!network) {
+      NET_LOG(ERROR) << "Unable to find eSIM network for request "
+                     << *uninstall_requests_.front();
+      CompleteCurrentRequest(UninstallESimResult::kRemoveServiceFailed);
+      return;
+    }
+
+    // Return success immediately for non-shill eSIM cellular networks since we
+    // don't know the actual shill service path. This stub non-shill service
+    // will be removed automatically when the eSIM profile list updates.
+    if (network->IsNonShillCellularNetwork()) {
+      CompleteCurrentRequest(UninstallESimResult::kSuccess);
+      return;
+    }
   }
 
-  // Return success immediately for non-shill eSIM cellular networks since we
-  // don't know the actual shill service path. This stub non-shill service will
-  // be removed automatically when the eSIM profile list updates.
-  if (network->IsNonShillCellularNetwork()) {
-    CompleteCurrentRequest(UninstallESimResult::kSuccess);
-    return;
-  }
-
+  NET_LOG(EVENT) << "Attempting to remove Shill service for network: "
+                 << network->path();
   network_configuration_handler_->RemoveConfiguration(
       network->path(), absl::nullopt,
       base::BindOnce(&CellularESimUninstallHandler::OnRemoveServiceSuccess,
@@ -315,6 +379,16 @@ void CellularESimUninstallHandler::AttemptRemoveShillService() {
 
 void CellularESimUninstallHandler::OnRemoveServiceSuccess() {
   DCHECK_EQ(state_, UninstallState::kRemovingShillService);
+  if (uninstall_requests_.front()->reset_euicc) {
+    // Wait for next network list update before removing the next shill service.
+    TransitionToUninstallState(UninstallState::kWaitingForNetworkListUpdate);
+    network_list_wait_timer_.Start(
+        FROM_HERE, kNetworkListWaitTimeout,
+        base::BindOnce(&CellularESimUninstallHandler::OnNetworkListWaitTimeout,
+                       weak_ptr_factory_.GetWeakPtr()));
+    return;
+  }
+
   CompleteCurrentRequest(UninstallESimResult::kSuccess);
 }
 
@@ -322,9 +396,16 @@ void CellularESimUninstallHandler::OnRemoveServiceFailure(
     const std::string& error_name,
     std::unique_ptr<base::DictionaryValue> error_data) {
   DCHECK_EQ(state_, UninstallState::kRemovingShillService);
-  NET_LOG(ERROR) << "Error removing service with ICCID "
-                 << GetIccidForCurrentRequest() << ". Error: " << error_name;
+  NET_LOG(ERROR) << "Error removing service for request "
+                 << *uninstall_requests_.front() << ". Error: " << error_name;
   CompleteCurrentRequest(UninstallESimResult::kRemoveServiceFailed);
+}
+
+void CellularESimUninstallHandler::OnNetworkListWaitTimeout() {
+  NET_LOG(ERROR)
+      << "Timedout waiting for network list update after removing service.";
+  TransitionToUninstallState(UninstallState::kRemovingShillService);
+  AttemptRemoveShillService();
 }
 
 NetworkStateHandler::NetworkStateList
@@ -345,14 +426,33 @@ CellularESimUninstallHandler::GetESimCellularNetworks() const {
   return network_list;
 }
 
-bool CellularESimUninstallHandler::HasQueuedRequest(
-    const std::string& iccid) const {
-  const auto iter = std::find_if(
-      uninstall_requests_.begin(), uninstall_requests_.end(),
-      [&](const std::unique_ptr<UninstallRequest>& uninstall_request) {
-        return uninstall_request->iccid == iccid;
-      });
-  return iter != uninstall_requests_.end();
+absl::optional<dbus::ObjectPath>
+CellularESimUninstallHandler::GetEnabledCellularESimProfilePath() {
+  for (const auto& esim_profile :
+       cellular_esim_profile_handler_->GetESimProfiles()) {
+    if (esim_profile.state() == CellularESimProfile::State::kActive) {
+      return esim_profile.path();
+    }
+  }
+  return absl::nullopt;
+}
+
+const NetworkState* CellularESimUninstallHandler::GetNextResetServiceToRemove()
+    const {
+  HermesEuiccClient::Properties* euicc_properties =
+      HermesEuiccClient::Get()->GetProperties(
+          *uninstall_requests_.front()->euicc_path);
+  const std::string& eid = euicc_properties->eid().value();
+  for (const NetworkState* network : GetESimCellularNetworks()) {
+    // Non Shill cellular services cannot be removed. They'll be automatically
+    // removed when eSIM profile list updates.
+    if (network->IsNonShillCellularNetwork()) {
+      continue;
+    }
+    if (network->eid() == eid)
+      return network;
+  }
+  return nullptr;
 }
 
 std::ostream& operator<<(
@@ -384,6 +484,21 @@ std::ostream& operator<<(
     case CellularESimUninstallHandler::UninstallState::kRemovingShillService:
       stream << "[Removing Shill Service]";
       break;
+    case CellularESimUninstallHandler::UninstallState::
+        kWaitingForNetworkListUpdate:
+      stream << "[Waiting for network list update]";
+      break;
+  }
+  return stream;
+}
+
+std::ostream& operator<<(
+    std::ostream& stream,
+    const CellularESimUninstallHandler::UninstallRequest& request) {
+  if (request.reset_euicc) {
+    stream << "(ResetEuicc)";
+  } else {
+    stream << "(ICCID: " << *request.iccid << ")";
   }
   return stream;
 }
